Metso_project/back-end/djangoreactproject/requirements.txt
django-cors-headers==2.4.0
djangorestframework==3.10.2
couchdb==0.10
Pillow==6.1.0
